Consolidate plutus script modules
